+++ src/draw/properties/primitive/quad.rs
@@ -11,7 +11,7 @@ use std::{iter, slice};
 
 /// Properties related to drawing a **Quad**.
 #[derive(Clone, Debug)]
-pub struct Quad<S = geom::DefaultScalar> {
+pub struct Quad<S = geom::scalar::Default> {
     quad: geom::Quad<Point3<S>>,
     spatial: spatial::Properties<S>,
     color: Option<Rgba>,