+++ src/draw/properties/primitive/tri.rs
@@ -11,7 +11,7 @@ use std::ops;
 
 /// Properties related to drawing a **Tri**.
 #[derive(Clone, Debug)]
-pub struct Tri<S = geom::DefaultScalar> {
+pub struct Tri<S = geom::scalar::Default> {
     tri: geom::Tri<Point3<S>>,
     spatial: spatial::Properties<S>,
     color: Option<Rgba>,