+++ src/geom/scalar.rs
@@ -1,0 +1,2 @@
+/// The default scalar type used for geometry throughout Nannou.
+pub type Default = f32;